Refactor MCP sandbox helpers
